Personal Website V2 (In Progress)
The second major project that i've had for my web development class. I did make another website for my first project, but it was so basic it wasn't worth throwing it up on the repo. I got to use Bootstrap and any other JS plugins I wanted to, to make this site.

The navbar came from startBootstrap and the parallax image came from a Google Search.